Refactor authorization handling by replacing AuthorizationService with AuthorityService and updating related method calls
