+++ src/main/resources/static/sql/schema.sql
@@ -53,3 +53,7 @@ ALTER TABLE member_authority ADD CONSTRAINT uq_member_authority_member_id_author
 ALTER TABLE authority ADD COLUMN systemId BIGINT REFERENCES system(id);
 ALTER TABLE authority DROP COLUMN systemId;
 ALTER TABLE authority ADD COLUMN system_id BIGINT REFERENCES system(id);
+
+ALTER TABLE authority DROP COLUMN value;
+ALTER TABLE authority ADD COLUMN value VARCHAR(31) NOT NULL ;
+ALTER TABLE authority ADD CONSTRAINT uq_authority_value_system_id UNIQUE (value, system_id);
